+++ src/js/renderers/twitch.js
@@ -55,9 +55,9 @@ const twitchApi = {
 			script.src = '//player.twitch.tv/js/embed/v1.js';
 
 			// Attach handlers for all browsers
-			script.onload = script.onreadystatechange = () => {
-				if (!done && (!twitchApi.readyState || twitchApi.readyState === undefined ||
-					twitchApi.readyState === "loaded" || twitchApi.readyState === "complete")) {
+			script.onload = script.onreadystatechange = function () {
+				if (!done && (!this.readyState || this.readyState === undefined ||
+					this.readyState === 'loaded' || this.readyState === 'complete')) {
 					done = true;
 					twitchApi.iFrameReady();
 					script.onload = script.onreadystatechange = null;