+++ changelog.md
@@ -8,6 +8,7 @@
 * Added `title` element for `mediaFiles` list in favor of `Source Chooser` plugin @rafa8626 
 * Added improvements and hover behavior to player slider (https://github.com/mediaelement/mediaelement/pull/2191) @ricking06 and @rafa8626
 * Added `babel-preset-env` to optimize bundles based on supported browsers @rafa8626
+* Fixed typos in `Vimeo` and `Twitch` renderers @rafa8626
 
 *4.0.7 (2017/04/25)*